notification-service/requirements.txt
channels
daphne
geoip2
py_eureka_client
aio-pika
psycopg2-binary
channels-redis
Django>=2.1.3
djangorestframework>=3.9.0
djangorestframework-gis
django-environ
django-leaflet
django-filter
django-cors-headers
python-dotenv
Pillow
requests
drf_yasg

